Add Triangle table and SQL query for triangle check
Create a Triangle table and insert sample data. Include a SQL query to determine if the segments can form a triangle , we just have to implement the property of triangle in which it says that  sum of any two sides of triangle is always bigger than that of the third side.